Change insert behavior to fit into best possible spot
The previous approach threw an error any time the insert location was
not available. With this fix, it fits it into the best spot available,
prioritized by targetIndex, followed by targetDepth. If the targetIndex
is reached, the node will be inserted as close to the targetDepth as
possible. If the targetIndex exceeds the size of the tree, add it to the
tree at the closest depth possible to the targetDepth.