fix: recalculate rowHeight() when isVirualized
When rendering rows with isVirualized, if rowHeight is a func, the
virtualized list was not recalculating the rowHeight.
Per the `react-virtualized` docs
[here](https://github.com/bvaughn/react-virtualized/blob/master/docs/Lis
t.md#recomputerowheights-index-number), `List.recomputeRowHeights()`
needs to be called.

This only fixes the dynamic rowHeight issue _after_ dragging a node is
finished.  An additional `List.recomputeRowHeights()` will need to be
called to fix the ‘preview’ while dragging a row (see issue below for
more details).

This partially fixes [Issue
264](https://github.com/fritz-c/react-sortable-tree/issues/264).